+++ swift_browser_ui_frontend/src/common/conv.js
@@ -227,7 +227,7 @@ export function makeGetObjectsMetaURL(project, container, objects) {
   );
 }
 
-export const taginputConfirmKeys = [",", ";", ":", ".", " ", "Tab", "Enter"];
+export const taginputConfirmKeys = ["Tab", "Enter", " ", ",", ";", ".", ":"];
 
 export function truncate(value, length) {
   if (!value) {